extract ListContent as a component
