+++ src/pages/Account/Center/Articles.less
@@ -1,7 +1,12 @@
+@import '~antd/lib/style/themes/default.less';
+
 .articleList {
   :global {
     .ant-list-item:first-child {
       padding-top: 0;
     }
   }
+}
+a.listItemMetaTitle {
+  color: @heading-color;
 }
